Fix the "vv4.0.0" on the main menu.
A little typo, Since we didn't that the new BLT was automatically adding an additional "v" on the main menu.
However, It didn't add a "v" on the mod menu SydneyHUD page, But that really a very small detail. Like the main menu double v.
You can see the double "v" on the bottom right on the picture. https://i.imgur.com/UafCAkf.jpg